- add app.ts file - refactor index.ts file to separate express server definition into app.ts and server startup login into index.ts
